+++ inc/ti/fn/fnsettype.h
@@ -28,7 +28,7 @@ static int do__f_set_type(ti_query_t * query, cleri_node_t * nd, ex_t * e)
         {
             ex_set(e, EX_VALUE_ERROR,
                     "name `%.*s` is reserved",
-                    (int) rname->n, (const char *) rname->data);
+                    rname->n, (const char *) rname->data);
             return e->nr;
         }
 
@@ -36,7 +36,7 @@ static int do__f_set_type(ti_query_t * query, cleri_node_t * nd, ex_t * e)
         {
             ex_set(e, EX_LOOKUP_ERROR,
                     "enum `%.*s` already exists",
-                    (int) rname->n, (const char *) rname->data);
+                    rname->n, (const char *) rname->data);
             return e->nr;
         }
